pyproject.toml
[project]
name = "databricks-personal"
version = "0.1.0"
description = "Personal Databricks workspace project"
requires-python = ">=3.10"
dependencies = [
    "databricks-sdk>=0.38.0",
    "python-dotenv>=1.0.0",
    "mlflow>=2.12.0",
    "transformers>=4.40.0",
    "torch>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.0.0",
    "pytest-cov>=7.0.0",
]

[project.scripts]
task = "tasks:cli"

# --------------------
# UV dependency groups
# --------------------
[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.14",
]

# --------------------
# ruff configuration
# --------------------
[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

# --------------------
# pytest configuration
# --------------------
[tool.pytest.ini_options]
addopts = "-ra -q --cov=src --cov-report=term-missing"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# --------------------
# coverage configuration
# --------------------
[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
